requirements.txt
apscheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
joblib==1.2.0
nltk==3.8.1
numpy==1.24.3
python-telegram-bot==13.13
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
scikit-learn==1.2.2
scipy==1.10.1
setuptools==67.7.1
six==1.16.0
threadpoolctl==3.1.0
tornado==6.3.1
tqdm==4.65.0
tzdata==2023.3
tzlocal==4.3

